First working api action for adding an element

+++ src/de/haw/mps/persistence/AbstractModel.java
@@ -4,72 +4,114 @@ import de.haw.mps.MpsLogger;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
+import javax.management.RuntimeErrorException;
 import java.io.Serializable;
 
 public abstract class AbstractModel<T extends Serializable> {
 
-    Class<T> tClass;
+    protected Class<T> tClass;
+
+    protected boolean hasActiveTransaction = false;
 
     public AbstractModel(Class<T> tClass) {
         this.tClass = tClass;
     }
 
-    public boolean add(T entity) {
+    public void startTransaction() {
+        if(!hasActiveTransaction) {
+            Session session = MpsSessionFactory.getcurrentSession();
+            session.beginTransaction();
+            hasActiveTransaction = true;
+        }
+    }
+
+    public void commitTransaction() throws WorkflowException {
+        if(!hasActiveTransaction) {
+            throw new WorkflowException("No transaction open.");
+        }
+
+        try {
+            Session session = MpsSessionFactory.getcurrentSession();
+            Transaction transaction = session.getTransaction();
+            transaction.commit();
+        } catch (Exception e) {
+            throw new WorkflowException(e);
+        }
+    }
+
+    public void rollbackTransaction() throws WorkflowException {
+        if(!hasActiveTransaction) {
+            throw new WorkflowException("No transaction open.");
+        }
+
+        try {
+            Session session = MpsSessionFactory.getcurrentSession();
+            Transaction transaction = session.getTransaction();
+            transaction.rollback();
+        } catch (Exception e) {
+            throw new WorkflowException(e);
+        }
+    }
+
+    public boolean add(T entity) throws WorkflowException {
         Session session = MpsSessionFactory.getcurrentSession();
         Transaction transaction = session.getTransaction();
         if(transaction == null || !transaction.isActive()) {
-            transaction = session.beginTransaction();
+            throw new WorkflowException("No transaction running while trying to insert new object.");
         }
 
         try {
             session.save(entity);
-            transaction.commit();
+            session.refresh(entity);
             return true;
         } catch (Exception e) {
-            transaction.rollback();
-            MpsLogger.handleException(e);
-            return false;
+            throw new WorkflowException(e);
         }
     }
 
-    public boolean update(T entity) {
+    public boolean update(T entity) throws WorkflowException {
         Session session = MpsSessionFactory.getcurrentSession();
-        Transaction transaction = session.beginTransaction();
+        Transaction transaction = session.getTransaction();
+        if(transaction == null || !transaction.isActive()) {
+            throw new WorkflowException("No transaction running while trying to update object.");
+        }
 
         try {
             session.update(entity);
             transaction.commit();
             return true;
         } catch (Exception e) {
-            transaction.rollback();
-            MpsLogger.handleException(e);
-            return false;
+            throw new WorkflowException(e);
         }
     }
 
-    public boolean delete(T entity) {
+    public boolean delete(T entity) throws WorkflowException {
         Session session = MpsSessionFactory.getcurrentSession();
-        Transaction transaction = session.beginTransaction();
+        Transaction transaction = session.getTransaction();
+        if(transaction == null || !transaction.isActive()) {
+            throw new WorkflowException("No transaction running while trying to delete object.");
+        }
 
         try {
             session.delete(entity);
             transaction.commit();
             return true;
         } catch (Exception e) {
-            transaction.rollback();
-            MpsLogger.handleException(e);
-            return false;
+            throw new WorkflowException(e);
         }
     }
 
-    public T get(Long id) {
+    public T get(Long id) throws WorkflowException {
         Session session = MpsSessionFactory.getcurrentSession();
+        Transaction transaction = session.getTransaction();
+        if(transaction == null || !transaction.isActive()) {
+            throw new WorkflowException("No transaction running while trying to get object.");
+        }
 
         try {
             return (T) session.get(tClass, id);
         } catch (Exception e) {
-            MpsLogger.handleException(e);
-            return null;
+            throw new WorkflowException(e);
         }
     }
 }
